+++ mtg_proxy_printer/settings.py
@@ -167,6 +167,9 @@ DEFAULT_SETTINGS["card-filter"] = {
     "hide-reversible-cards": "False",
     "hide-art-series-cards": "False",
     "hide-universes-beyond-cards": "False",
+    "hide-low-resolution-cards": "False",
+    "hide-textless-cards": "False",
+    "hide-full-art-cards": "False",
     "hidden-sets": "",
 }